+++ praw/__init__.py
@@ -927,16 +927,15 @@ class UnauthenticatedReddit(BaseReddit):
                 raise
         return result
 
-    def search(self, query, subreddit=None, sort=None, syntax=None, time=None,
-               *args, **kwargs):
+    def search(self, query, subreddit=None, sort=None, syntax=None,
+               period=None, *args, **kwargs):
         """Return a generator for submissions that match the search query.
 
         :param query: The query string to search for.
         :param subreddit: Limit search results to the subreddit if provided.
         :param sort: The sort order of the results.
         :param syntax: The syntax of the search query.
-        :param time: The time window of the results. Similar to the time window
-            of subreddit listings.
+        :param period: The time period of the results.
 
         The additional parameters are passed directly into
         :meth:`.get_content`. Note: the `url` and `param` parameters cannot be
@@ -951,8 +950,8 @@ class UnauthenticatedReddit(BaseReddit):
             params['sort'] = sort
         if syntax:
             params['syntax'] = syntax
-        if time:
-            params['t'] = time
+        if period:
+            params['t'] = period
         if subreddit:
             params['restrict_sr'] = 'on'
             url = self.config['search'] % subreddit